+++ .gitignore
@@ -1,9 +1,11 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
-/node_modules
+/frontEnd/node_modules
 /.pnp
 .pnp.js
+
+
 
 # testing
 /coverage
@@ -21,3 +23,5 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+*.pyc
+*.pyc
