--- Front End/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "attempt-1",
-  "version": "1.0.0",
-  "description": "Code Club App",
-  "main": "main.js",
-  "scripts": {
-    "start": "electron ."
-  },
-  "author": "Code Club",
-  "license": "ISC",
-  "dependencies": {
-    "child_process": "^1.0.2",
-    "python-shell": "^2.0.3"
-  },
-  "devDependencies": {
-    "electron": "^11.2.1"
-  }
-}
